+++ bert/providers/opinions.py
@@ -1,3 +1,4 @@
+from arekit.common.experiment.data_type import DataType
 from arekit.common.experiment.formats.base import BaseExperiment
 from arekit.common.experiment.opinions import extract_text_opinions
 from arekit.common.labels.base import Label
@@ -16,7 +17,7 @@ class OpinionProvider(object):
     """
 
     def __init__(self, data_type, text_opinions, parsed_news_collection):
-        assert(isinstance(data_type, unicode))
+        assert(isinstance(data_type, DataType))
         assert(isinstance(text_opinions, LinkedTextOpinionCollection))
         assert(isinstance(parsed_news_collection, ParsedNewsCollection))
         self.__text_opinions = text_opinions