+++ common/experiment/neutral/annot/base.py
@@ -43,7 +43,7 @@ class BaseNeutralAnnotator(object):
 
     # TODO. Supported is considered by experiment.
     def __iter_all_doc_ids(self):
-        for data_type in DataType.iter_supported():
+        for data_type in self._DocOps.iter_suppoted_data_types():
             for doc_id in self._DocOps.iter_news_indices(data_type):
                 yield doc_id